requirements.txt
pygame
soco
python-decouple
rpi-backlight